Adds the HLL solver
This commit also changes the policy of a Riemann solver. The input
states must contain valid (matching) conserved and primitive
values. Fluxes and eigenvalues will then be obtained as needed
within. The output state is guarenteed to have a conserved value and a
flux. The flux in the case of HLL is not the flux derived from those
values, but rather from the HLL formula directly, which satisfies the
jump conditions. After the call is executed the user is free to obtain
the primitives by calling c2p on the output state, although primitives
might already have been calculated, as in the case of the exact solver.

That policy may change, but until it does, this exact paragraph is
contained at the top of riemann.c.