Updates Cython code to account for changes in the C API
This commit also has the side effect of fixing a bug in the C code,
where incomplete states were not detected due to using a '&' instead of
a '&&' in the _nrhyd_update function. An error code for incomplete
states is also introduced. That bug fix will now be merged into master
along with the first appearance of the Python wrappers.